# auth-service/src/main/resources/openapi.json
{
  "openapi": "3.1.0",
  "info": {"title": "CityBuddy auth identity API", "version": "0.0.1"},
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginRequest"}}}
        },
        "responses": {"200": {"description": "RS256 direct-user bearer token", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenResponse"}}}}, "401": {"description": "Invalid credentials, disabled principal, or production evaluation context"}}
      }
    },
    "/auth/jwks": {
      "get": {
        "operationId": "jwks",
        "parameters": [{"name": "X-Eval-Sandbox-Id", "in": "header", "required": false, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Current and overlapping public RSA keys only", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/JwksResponse"}}}}, "401": {"description": "Production evaluation context is rejected"}}
      }
    },
    "/auth/token/exchange": {
      "post": {
        "operationId": "exchangeAgentObo",
        "description": "The authenticated service fixes the actor and permission policy. agent-service requires support:session:create and cannot request merchant:* or shopping:* scopes. merchant-agent requires merchant:session:create and permits only merchant:read, merchant:price:prepare, merchant:price:read, or merchant:price:cancel. shopping-agent requires shopping:session:create and permits only shopping:orders:read, shopping:cart:read, shopping:cart:write, shopping:profile:read, or refund:create; refunds retain their existing exact scope and use the shopping-agent actor. Both service and deployment scope grants must allow the requested scope. Permissions come from the signed direct token, never the request body. The service owns and validates the session-to-user binding before asserting it here; Auth verifies the asserted user against the direct token. No merchant:price:apply OBO scope is issued.",
        "security": [{"serviceBasic": []}],
        "parameters": [{"name": "X-User-Authorization", "in": "header", "required": true, "schema": {"type": "string"}}, {"name": "X-Eval-Sandbox-Id", "in": "header", "required": false, "schema": {"$ref": "#/components/schemas/SandboxId"}, "description": "Required only for an evaluation direct-user token; must equal its signed sandbox claim."}],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ExchangeRequest"}}}
        },
        "responses": {"200": {"description": "Short-lived exact-scope agent OBO", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenResponse"}}}}, "401": {"description": "Invalid direct user, service credential, or production evaluation context"}, "403": {"description": "Invalid session assertion, principal state, or scope"}}
      }
    },
    "/internal/eval/test-principals/provision": {
      "post": {
        "operationId": "provisionEvaluationTestPrincipal",
        "x-citybuddy-profile": "evaluation",
        "security": [{"commerceEvaluationBasic": []}],
        "parameters": [{"name": "Idempotency-Key", "in": "header", "required": true, "schema": {"$ref": "#/components/schemas/BoundedIdentifier128"}}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EvaluationProvisionRequest"}}}},
        "responses": {"200": {"description": "Opaque TTL-bound evaluation test-principal handle", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EvaluationProvisionResponse"}}}}, "400": {"description": "Malformed or out-of-bounds provisioning intent"}, "401": {"description": "Invalid dedicated commerce-service credential"}, "409": {"description": "Conflicting idempotency or sandbox/case reuse"}}
      }
    },
    "/internal/eval/test-principals/{handle}/revoke": {
      "post": {
        "operationId": "revokeEvaluationTestPrincipal",
        "x-citybuddy-profile": "evaluation",
        "security": [{"commerceEvaluationBasic": []}],
        "parameters": [{"name": "handle", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/OpaqueEvaluationHandle"}}, {"name": "Idempotency-Key", "in": "header", "required": true, "schema": {"$ref": "#/components/schemas/BoundedIdentifier128"}}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EvaluationRevokeRequest"}}}},
        "responses": {"200": {"description": "Idempotent revoked state", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EvaluationRevokeResponse"}}}}, "400": {"description": "Malformed revocation intent"}, "401": {"description": "Invalid dedicated commerce-service credential"}, "404": {"description": "Unknown or mismatched private binding"}, "409": {"description": "Conflicting revocation idempotency reuse"}}
      }
    },
    "/auth/eval/test-token": {
      "post": {
        "operationId": "issueEvaluationTestToken",
        "x-citybuddy-profile": "evaluation",
        "security": [{"evaluatorBasic": []}],
        "parameters": [{"name": "X-Eval-Sandbox-Id", "in": "header", "required": true, "schema": {"$ref": "#/components/schemas/SandboxId"}}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EvaluationTestTokenRequest"}}}},
        "responses": {"200": {"description": "RS256 sandbox-bound direct-user evaluation token", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenResponse"}}}}, "400": {"description": "Malformed handle or sandbox"}, "401": {"description": "Invalid independent evaluator credential, handle, lifecycle state, or sandbox binding"}}
      }
    }
  },
  "components": {
    "securitySchemes": {"serviceBasic": {"type": "http", "scheme": "basic"}, "commerceEvaluationBasic": {"type": "http", "scheme": "basic", "description": "Dedicated commerce-service evaluation management identity"}, "evaluatorBasic": {"type": "http", "scheme": "basic", "description": "Independent evaluation test-token issuer identity"}},
    "schemas": {
      "SandboxId": {"type": "string", "minLength": 1, "maxLength": 64, "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$"},
      "BoundedIdentifier128": {"type": "string", "minLength": 1, "maxLength": 128, "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$"},
      "OpaqueEvaluationHandle": {"type": "string", "minLength": 43, "maxLength": 43, "pattern": "^[A-Za-z0-9_-]{43}$"},
      "EvaluationProvisionRequest": {"type": "object", "additionalProperties": false, "required": ["sandboxId", "caseCorrelation", "testUserLabel", "ttlSeconds"], "properties": {"sandboxId": {"$ref": "#/components/schemas/SandboxId"}, "caseCorrelation": {"$ref": "#/components/schemas/BoundedIdentifier128"}, "testUserLabel": {"$ref": "#/components/schemas/BoundedIdentifier128"}, "ttlSeconds": {"type": "integer", "minimum": 60, "maximum": 3600}}},
      "EvaluationProvisionResponse": {"type": "object", "additionalProperties": false, "required": ["handle", "expiresAt"], "properties": {"handle": {"$ref": "#/components/schemas/OpaqueEvaluationHandle"}, "expiresAt": {"type": "string", "format": "date-time", "readOnly": true}}},
      "EvaluationRevokeRequest": {"type": "object", "additionalProperties": false, "required": ["sandboxId", "caseCorrelation"], "properties": {"sandboxId": {"$ref": "#/components/schemas/SandboxId"}, "caseCorrelation": {"$ref": "#/components/schemas/BoundedIdentifier128"}}},
      "EvaluationRevokeResponse": {"type": "object", "additionalProperties": false, "required": ["handle", "state"], "properties": {"handle": {"$ref": "#/components/schemas/OpaqueEvaluationHandle"}, "state": {"const": "REVOKED", "readOnly": true}}},
      "EvaluationTestTokenRequest": {"type": "object", "additionalProperties": false, "required": ["handle"], "properties": {"handle": {"$ref": "#/components/schemas/OpaqueEvaluationHandle"}}},
      "LoginRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["loginIdentifier", "password"],
        "properties": {"loginIdentifier": {"type": "string"}, "password": {"type": "string", "format": "password"}}
      },
      "ExchangeRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["sessionId", "userSubject", "scope"],
        "properties": {"sessionId": {"type": "string", "minLength": 1}, "userSubject": {"type": "string", "minLength": 1}, "scope": {"type": "string", "minLength": 1, "pattern": "^[^*\\s]+$"}}
      },
      "TokenResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["accessToken", "tokenType", "expiresIn"],
        "properties": {"accessToken": {"type": "string"}, "tokenType": {"const": "Bearer"}, "expiresIn": {"type": "integer", "minimum": 1}}
      },
      "JwksResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["keys"],
        "properties": {"keys": {"type": "array", "items": {"$ref": "#/components/schemas/PublicRsaJwk"}}}
      },
      "PublicRsaJwk": {
        "type": "object",
        "additionalProperties": false,
        "required": ["kty", "kid", "alg", "n", "e"],
        "properties": {"kty": {"const": "RSA"}, "kid": {"type": "string"}, "alg": {"const": "RS256"}, "use": {"type": "string"}, "n": {"type": "string"}, "e": {"type": "string"}}
      }
    }
  }
}